refactor: moved getMartinTileJson
