+++ sites/geohub/src/routes/api/vector/[source]/metadata.json/+server.ts
@@ -2,7 +2,7 @@ import type { RequestHandler } from './$types'
 import type { TileJson } from '$lib/types/TileJson'
 import type { VectorTileMetadata } from '$lib/types/VectorTileMetadata'
 import { getStaticPbfMetadataJson } from '$lib/server/helpers/getStaticPbfMetadataJson'
-import { getMartinTileJson } from '$lib/helper/getMartinTileJson'
+import { getMartinTileJson } from '$lib/server/helpers/getMartinTileJson'
 import { generateMetadataJson } from '$lib/server/helpers/generateMetadataJson'
 import { getPgtileservTileJson } from '$lib/helper/getPgtileservTileJson'
 import { error } from '@sveltejs/kit'
